working in ui and split components

+++ components/Header.tsx
@@ -1,8 +1,9 @@
 import Toolbar from '@mui/material/Toolbar';
 import IconButton from '@mui/material/IconButton';
 import SearchIcon from '@mui/icons-material/Search';
-import {useState} from 'react';
+import { useState } from 'react';
 import Button from '@mui/material/Button';
+import Box from '@mui/material/Box';
 import TextField from '@mui/material/TextField';
 import Dialog from '@mui/material/Dialog';
 import DialogActions from '@mui/material/DialogActions';
@@ -12,8 +13,8 @@ import DialogTitle from '@mui/material/DialogTitle';
 import Link from 'next/link';
 import useGetUser from '../hooks/useGetUser';
 import { Logout } from '../api';
-import Router, { useRouter } from 'next/router';
-import SideBarHead from './SideBarHead';
+import { useRouter } from 'next/router';
+import Swal from 'sweetalert2';
 
 interface ISearchProps {
   open: boolean;
@@ -67,38 +68,50 @@ export default function Header() {
   const [search, setSearch] = useState("");
   const [user] = useGetUser();
   const router = useRouter()
-  
+
   const handelLogout = async () => {
-    await Logout();
-    localStorage.clear();
-    await router.push("/login")
-    router.reload()
+    Swal.fire({
+      title: 'Are you sure you want to Logout',
+      icon: "warning",
+      showCancelButton: true,
+      showConfirmButton: true
+    })
+      .then(async (res) => {
+        if (!res.value) return;
+        await Logout();
+        localStorage.clear();
+        await router.push("/login")
+        router.reload()
+      })
+
+  }
+
+  const handelSingUp = async () => {
+    await router.push("/sing-up")
   }
 
   return (
     <header>
       <Search open={open} setOpen={setOpen} search={search} setSearch={setSearch} />
-      <Toolbar className="flex justify-between" sx={{ borderBottom: 1, borderColor: 'divider' }}> 
-      <SideBarHead />
+      <Toolbar className="min-w-[95vw]" sx={{ borderBottom: 1, borderColor: 'divider' }}>
 
-        <div>
-           <IconButton className="mr-2" onClick={() => setOpen(true)}>
+        <div className="inline-flex w-full">
+          <IconButton className="mr-2" onClick={() => setOpen(true)}>
             <SearchIcon />
           </IconButton>
-          {user !== null ? (
-            <Button onClick={handelLogout} variant="outlined" size="small">
-              logout
-            </Button>
-          ) : (
-            <Link className="no-underline" href="/sing-up">
-              <Button variant="outlined" size="small">
-                Sign up
-              </Button>
+
+          <Box className="inline-flex w-full justify-between">
+            {user !== null
+              ? <Button onClick={handelLogout} variant="outlined" size="small"> Logout </Button>
+              : <Button onClick={handelSingUp} variant="outlined" size="small"> Sign up </Button>}
+
+            <Link href="/admin/dashboard">
+              <Button onClick={handelSingUp} variant="outlined" size="small"> DashBoard </Button>
             </Link>
-          )} 
+          </Box>
         </div>
 
-       </Toolbar> 
+      </Toolbar>
     </header>
   );
 }
